clean warnings on msvc

+++ custom_ops.cpp
@@ -21,9 +21,9 @@
 
 static float qwen_get_ntk_alpha(int true_seq_len, int seq_length)
 {
-    float context_value = log2f(true_seq_len / seq_length) + 1;
+    float context_value = log2f((float)true_seq_len / (float)seq_length) + 1;
     float ntk_alpha = powf(2, ceil(context_value)) - 1;
-    return MAX(ntk_alpha, 1.0);
+    return MAX(ntk_alpha, 1.0f);
 }
 
 static void ggml_compute_forward_mat_scale_f32(struct ggml_tensor * dst , const struct ggml_tensor * a, const struct ggml_tensor * b, int ith, int nth, void * userdata)
@@ -33,7 +33,7 @@ static void ggml_compute_forward_mat_scale_f32(struct ggml_tensor * dst , const 
 
     GGML_TENSOR_BINARY_OP_LOCALS
 
-    const int nr = ggml_nrows(dst);
+    const int nr = (int)ggml_nrows(dst);
 
     // rows per thread
     const int dr = (nr + nth - 1)/nth;
@@ -90,7 +90,7 @@ static void ggml_compute_forward_ntk_dynamic_rope_f32(struct ggml_tensor * dst ,
 
     int n_dims = data->rope_dim;
 
-    const int nr = ggml_nrows(dst);
+    const int nr = (int)ggml_nrows(dst);
 
     GGML_ASSERT(n_dims <= ne0);
     GGML_ASSERT(n_dims % 2 == 0);
@@ -115,15 +115,15 @@ static void ggml_compute_forward_ntk_dynamic_rope_f32(struct ggml_tensor * dst ,
                 if (ir   > ir1) break;
 
                 const float ntk_alpha = qwen_get_ntk_alpha(p, data->seq_length);
-                const float base = data->freq_base * powf(ntk_alpha, n_dims / (n_dims - 2.0f));
-                const float inv_freq_scale = powf(base, -2.0f / n_dims);
+                const float base = data->freq_base * powf(ntk_alpha, (float)n_dims / ((float)n_dims - 2.0f));
+                const float inv_freq_scale = powf(base, -2.0f / (float)n_dims);
                 float inv_freq = 1.0f;
 
                 for (int64_t ic = 0; ic < ne0; ic += 2, inv_freq *= inv_freq_scale) {
                     if (ic < n_dims) {
                         const int64_t ib = 0;
 
-                        float theta = p * inv_freq;
+                        float theta = (float)p * inv_freq;
                         float cos_theta = cosf(theta);
                         float sin_theta = sinf(theta);
 
@@ -163,7 +163,7 @@ static void ggml_compute_forward_ntk_dynamic_rope_f16(struct ggml_tensor * dst ,
 
     int n_dims = data->rope_dim;
 
-    const int nr = ggml_nrows(dst);
+    const int nr = (int)ggml_nrows(dst);
 
     GGML_ASSERT(n_dims <= ne0);
     GGML_ASSERT(n_dims % 2 == 0);
@@ -188,15 +188,15 @@ static void ggml_compute_forward_ntk_dynamic_rope_f16(struct ggml_tensor * dst ,
                 if (ir   > ir1) break;
 
                 const float ntk_alpha = qwen_get_ntk_alpha(p, data->seq_length);
-                const float base = data->freq_base * powf(ntk_alpha, n_dims / (n_dims - 2.0f));
-                const float inv_freq_scale = powf(base, -2.0f / n_dims);
+                const float base = data->freq_base * powf(ntk_alpha, (float)n_dims / ((float)n_dims - 2.0f));
+                const float inv_freq_scale = powf(base, -2.0f / (float)n_dims);
                 float inv_freq = 1.0f;
 
                 for (int64_t ic = 0; ic < ne0; ic += 2, inv_freq *= inv_freq_scale) {
                     if (ic < n_dims) {
                         const int64_t ib = 0;
 
-                        float theta = p * inv_freq;
+                        float theta = (float)p * inv_freq;
                         float cos_theta = cosf(theta);
                         float sin_theta = sinf(theta);
 
@@ -252,7 +252,7 @@ static void ggml_compute_forward_ntk_mix_rope_f32(struct ggml_tensor * dst , con
 
     int n_dims = data->rope_dim;
 
-    const int nr = ggml_nrows(dst);
+    const int nr = (int)ggml_nrows(dst);
 
     GGML_ASSERT(n_dims <= ne0);
     GGML_ASSERT(n_dims % 2 == 0);
@@ -308,7 +308,7 @@ static void ggml_compute_forward_ntk_mix_rope_f16(struct ggml_tensor * dst , con
 
     int n_dims = data->rope_dim;
 
-    const int nr = ggml_nrows(dst);
+    const int nr = (int)ggml_nrows(dst);
 
     GGML_ASSERT(n_dims <= ne0);
     GGML_ASSERT(n_dims % 2 == 0);
@@ -373,11 +373,11 @@ static void build_ntk_mixed_inv_freq(int dim, std::vector<float> &inv_freq,
     int max_position_embeddings = 2048, float base = 10000.0, float k = 16, float b = 0.3)
 {
     inv_freq.clear();
-    float a = log(k) / powf(dim / 2, b);
+    float a = log(k) / powf((float)dim / 2, b);
     inv_freq.reserve(dim / 2);
     for (int i = 0; i < dim; i += 2)
     {
-        float v = 1.0f / powf(base, (float)i / dim) / expf(a * powf(i / 2 + 1, b));
+        float v = 1.0f / powf(base, (float)i / (float)dim) / expf(a * powf((float)i / 2 + 1, b));
         inv_freq.push_back(v);
     }
 }
@@ -390,7 +390,7 @@ static void ggml_compute_forward_simple_norm_f32(struct ggml_tensor * dst , cons
     GGML_TENSOR_UNARY_OP_LOCALS
 
     // TODO: make this tunable
-    float eps = 1e-6;
+    float eps = 1e-6f;
 
     GGML_ASSERT(eps > 0.0f);
 
@@ -407,7 +407,7 @@ static void ggml_compute_forward_simple_norm_f32(struct ggml_tensor * dst , cons
 
                 float * y = (float *) ((char *) dst->data + i01*nb1 + i02*nb2 + i03*nb3);
 
-                const float scale = 1.0f/sqrtf(sum + eps);
+                const float scale = 1.0f / sqrtf((float)sum + eps);
 
                 for (int64_t i = 0; i < ne00; i++)
                     y[i] = x[i] * scale;
@@ -437,7 +437,7 @@ static void ggml_compute_forward_sigmoid_f32(struct ggml_tensor * dst , const st
     GGML_TENSOR_UNARY_OP_LOCALS
 
     // TODO: make this tunable
-    float eps = 1e-6;
+    float eps = 1e-6f;
 
     GGML_ASSERT(eps > 0.0f);
